Add new python feature to make Python support conditional (#748)
