+++ crates/arroyo-udf/arroyo-udf-python/Cargo.toml
@@ -3,11 +3,14 @@ name = "arroyo-udf-python"
 version = "0.2.0"
 edition = "2021"
 
+[features]
+python-enabled = ["pyo3"]
+
 [dependencies]
 arroyo-udf-common = { path = "../arroyo-udf-common" }
 arrow = { workspace = true, features = ["ffi"] }
 datafusion = { workspace = true }
-pyo3 = { version = "0.21"}
+pyo3 = { version = "0.21", optional = true}
 anyhow = "1"
 tokio = { version = "1", features = ["full"] }
 itertools = "0.13.0"
