+++ crates/arroyo/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [features]
 profiling = ["tikv-jemallocator/profiling"]
+python = ["arroyo-udf-python/python-enabled"]
 
 [dependencies]
 arroyo-types = { path ="../arroyo-types" }
@@ -18,6 +19,7 @@ arroyo-node = { path = "../arroyo-node" }
 arroyo-rpc = { path = "../arroyo-rpc" }
 arroyo-openapi = { path ="../arroyo-openapi" }
 arroyo-storage = { path = "../arroyo-storage" }
+arroyo-udf-python = { path = "../arroyo-udf/arroyo-udf-python" }
 
 clap = { version = "4", features = ["derive"] }
 tokio = { version = "1", features = ["full"] }
